Renamed unit_test.py to preliminary_tests.py; and added more comments
